+++ changelog.md
@@ -3,7 +3,9 @@
 ## v3.0.9
 
 - 暂时禁用文档树缩进线
+- 在通过标题、图标展开收起文档树时，在可点击打开文档处显示特殊光标。需在主题设置中启用
 - Temporarily disable document tree indent line
+- When expanding and collapsing the document tree by clicking the title or icon, show a special cursor at where you can open the document. Need to enable in theme settings
 
 ## v3.0.8